app/src/test/java/com/fongmi/android/tv/player/EpgUtilTest.java
package com.fongmi.android.tv.test;

import com.fongmi.android.tv.bean.Epg;
import com.fongmi.android.tv.player.EpgUtil;
import com.github.catvod.net.OkHttp;
import org.testng.Assert;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import java.util.Map;

public class EpgUtilTest {

    private EpgUtil epgUtil;
    private static final String TEST_EPG_URL = "https://epg.erw.cc/cc.xml";

    @BeforeMethod
    public void setUp() {
        epgUtil = new EpgUtil();
    }

    @Test
    public void testParseEpgFromXmlSource() {
        // Add assertions to validate the result of parseEpgFromXmlSource()
    }

    @Test
    public void testParseEpgFromXml() {
        Map<String, Epg> epgMap = epgUtil.parseEpgFromXmlSource(TEST_EPG_URL);

        Epg epg1 = epgUtil.getEpgByChannelName("CCTV1");
        Assert.assertNotNull(epg1, "Epg for channel 'CCTV1' should not be null");

        Epg epg2 = epgUtil.getEpgByChannelId("85");
        Assert.assertNotNull(epg2, "Epg for channel '85' should not be null");

        Epg epg3 = epgUtil.getEpgByChannelId("10086");
        Assert.assertNull(epg3, "Epg for channel '10086' should be null"); // Assuming it doesn't exist

        // Remove System.out.println() statements in unit tests, they clutter the output
    }

    @Test
    public void testGetEpgMap() {
        // Add assertions to validate the result of getEpgMap()
    }

    @Test
    public void testGetEpgByChannelName() {
        // Add specific test cases for getEpgByChannelName()
    }

    @Test
    public void testGetEpgByChannelId() {
        // Add specific test cases for getEpgByChannelId()
    }

    @Test
    public void testDownload() {
        // Add test cases for download functionality
    }

    @Test
    public void testProgress() {
        // Add test cases for progress tracking
    }

    @Test
    public void testError() {
        // Add test cases for error scenarios
    }

    @Test
    public void testSuccess() {
        // Add test cases for successful scenarios
    }
}
